Move plugins-related extension points and BaseAction to new picard.extension_points.item_actions
- it breaks plugins of course
